Put Function for Cart working and Added several js pages for Checkout
CartScreen.js has the ability to change the amount in edit form

Checkout.js currently has a mini router to switch between Delivery -> Payment -> Confirm. This is most likely going to be  modified to show a green tick to show which process of checkout is done.

Delivery.js has a form for Address as of now

27/10/2021
Daniel W 45583552